Flask-VanillaJs2/server/requirements.py3.10._.txt
# 
# python 3.10 compliant
#

Flask==3.1.0
environs==14.1.1
# google client ( for google services )
google-api-python-client==2.154.0
# google auth
google-auth-oauthlib==1.2.1
Flask-Session==0.8.0
redis